GluttonyCombo v1.0.4.84 - sync upstream WrathCombo (DNC/NIN/opener refinements)
Merged wrathcombo/main aede233c6..b4e7f972f (7 commits): DNC opener/prepull tuning, NIN adjusted-action mudra fix, opener skip safety, MCH/RDM/SAM tweaks. Per-file 3-way, 0 conflicts, all fork divergences preserved. Changelog plumbed across all 5 copies. Build 0 errors.

+++ src/GluttonyCombo/GluttonyCombo/Combos/PvE/SAM/SAM_Helper.cs
@@ -583,7 +583,7 @@ internal partial class SAM
             ([2], 11, () => !TargetNeedsPositionals())
         ];
 
-        public override List<(int[] Steps, Func<int> HoldDelay)> PrepullDelays { get; set; } =
+        public override List<(int[] Steps, Func<float> HoldDelay)> PrepullDelays { get; set; } =
         [
             ([2], () => SAM_Opener_PrePullDelay)
         ];
@@ -636,7 +636,7 @@ internal partial class SAM
             ([2], 11, () => !TargetNeedsPositionals())
         ];
 
-        public override List<(int[] Steps, Func<int> HoldDelay)> PrepullDelays { get; set; } =
+        public override List<(int[] Steps, Func<float> HoldDelay)> PrepullDelays { get; set; } =
         [
             ([2], () => SAM_Opener_PrePullDelay)
         ];
@@ -691,7 +691,7 @@ internal partial class SAM
             ([2], 11, () => !TargetNeedsPositionals())
         ];
 
-        public override List<(int[] Steps, Func<int> HoldDelay)> PrepullDelays { get; set; } =
+        public override List<(int[] Steps, Func<float> HoldDelay)> PrepullDelays { get; set; } =
         [
             ([2], () => SAM_Opener_PrePullDelay)
         ];
@@ -761,7 +761,7 @@ internal partial class SAM
             ([14], () => SenCount is not 1 && !(SenCount is 2 && JustUsed(Gekko)))
         ];
 
-        public override List<(int[] Steps, Func<int> HoldDelay)> PrepullDelays { get; set; } =
+        public override List<(int[] Steps, Func<float> HoldDelay)> PrepullDelays { get; set; } =
         [
             ([2], () => SAM_Opener_PrePullDelay)
         ];
